+++ test/IRGen/Inputs/cdecl_implementation.h
@@ -1,0 +1,2 @@
+extern void implFunc(int param);
+extern void implFuncCName(int param) __asm__("_implFuncAsmName");
