+++ test/IRGen/Inputs/objc_implementation.h
@@ -32,10 +32,6 @@
 
 @end
 
-extern void implFunc(int param);
-extern void implFuncCName(int param) __asm__("_implFuncAsmName");
-
-
 @interface NoImplClass
 
 - (void)noImplMethod:(int)param;
