+++ test/Interpreter/objc_implementation_objc_client.m
@@ -6,7 +6,7 @@
 // RUN: %empty-directory(%t-frameworks/objc_implementation.framework/Headers)
 // RUN: cp %S/Inputs/objc_implementation.modulemap %t-frameworks/objc_implementation.framework/Modules/module.modulemap
 // RUN: cp %S/Inputs/objc_implementation.h %t-frameworks/objc_implementation.framework/Headers
-// RUN: %target-build-swift-dylib(%t-frameworks/objc_implementation.framework/objc_implementation) -emit-module-path %t-frameworks/objc_implementation.framework/Modules/objc_implementation.swiftmodule/%module-target-triple.swiftmodule -module-name objc_implementation -F %t-frameworks -import-underlying-module -Xlinker -install_name -Xlinker %t-frameworks/objc_implementation.framework/objc_implementation %S/objc_implementation.swift -enable-experimental-feature CImplementation -target %target-stable-abi-triple
+// RUN: %target-build-swift-dylib(%t-frameworks/objc_implementation.framework/objc_implementation) -emit-module-path %t-frameworks/objc_implementation.framework/Modules/objc_implementation.swiftmodule/%module-target-triple.swiftmodule -module-name objc_implementation -F %t-frameworks -import-underlying-module -Xlinker -install_name -Xlinker %t-frameworks/objc_implementation.framework/objc_implementation %S/objc_implementation.swift -target %target-stable-abi-triple
 
 //
 // Execute this file
@@ -72,7 +72,6 @@ static void printString(Class class, NSString *property, NSString *str) {
 
 int main() {
   [ImplClass runTests];
-  // CHECK: implFunc(1989)
   // CHECK-LABEL: *** ImplClass init ***
   // CHECK: ImplClass.init()
   // CHECK-RESILIENCE-LABEL: *** ImplClassWithResilientStoredProperty #1 ***
