+++ test/PrintAsObjC/Inputs/custom-modules/cdecl_implementation/cdecl_implementation.h
@@ -1,0 +1,1 @@
+void CImplFunc(void);
