+++ test/PrintAsObjC/Inputs/custom-modules/objc_implementation/objc_implementation.h
@@ -9,5 +9,3 @@
 @interface ObjCClass2 : NSObject
 
 @end
-
-void CImplFunc(void);
